fix: increase frame window for consumption computation
my new linky sends frames every second (previous one was every 2 seconds).
The frame window used to compute consumption was now filled in 40 secs instead of 80.
In case of low power consumption (less that 2 Wh in 40 sec) the consumption was calculated
using the oldest frame of the window and did not reflect real values.
This commit increases the window size in order to avoid this behaviour.